// packages/core/src/theme.d.ts
export declare const theme: {
    color: {
        background: string;
        primary: string;
        secondary: string;
        destructive: string;
        ghost: string;
        link: string;
        blue: string;
        red: string;
        green: string;
        yellow: string;
        orange: string;
        white: string;
        black: string;
        off10: string;
        off20: string;
        off30: string;
        off40: string;
        off50: string;
        alert: string;
        warning: string;
        title: string;
        label: string;
        textPrimary: string;
        textSecondary: string;
        textGhost: string;
        textLink: string;
        borderPrimary: string;
        borderSecondary: string;
        borderDestructive: string;
        borderGhost: string;
        true: string;
        false: string;
        muted: string;
        activeText: string;
        text: string;
    };
    size: {
        headtitle: number;
        title: number;
        label: number;
        sublabel: number;
        small: number;
    };
    font: {
        black: string;
        bold: string;
        medium: string;
        book: string;
    };
};
export default theme;
//# sourceMappingURL=theme.d.ts.map